Updated user/update api with redis cache integration

+++ src/main/java/com/tanmoy/springboottest/controller/UserController.java
@@ -6,6 +6,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.Objects;
+
 import static org.springframework.http.ResponseEntity.ok;
 
 @RestController
@@ -20,8 +22,11 @@ public class UserController {
         return ok(userService.saveUser(userDto));
     }
 
-    @PostMapping(value = "/update")
+    @PutMapping(value = "/update")
     public ResponseEntity<Object> update(@RequestBody UserDto userDto) {
+        if (Objects.isNull(userDto.getUserId())) {
+            return ok("Please provide user id to update user");
+        }
         return ok(userService.updateUser(userDto));
     }
 
@@ -37,6 +42,9 @@ public class UserController {
 
     @GetMapping(value = "/delete/{id}")
     public ResponseEntity<Object> deleteUser(@PathVariable Long id) {
+        if (Objects.isNull(id)) {
+            return ok("Please provide user id to delete user");
+        }
         boolean isDeleted = userService.deleteUser(id);
         return ok(isDeleted ? "User has been deleted successfully" : "Failed to delete user");
     }
